+++ synthesizer/src/vm/tests/mod.rs
@@ -41,4 +41,7 @@ mod test_v16;
 mod test_v17;
 
 #[cfg(feature = "test")]
+mod test_v18;
+
+#[cfg(feature = "test")]
 use super::*;
